EXLM-1305: per PR review comments. Make browse-filter-utils lighter.
